pyproject.toml
[project]
name = "gpt2"
version = "0.1.0"
dependencies = [
    "hydra-core",
    "tiktoken",
    "torch",
    "numpy",
    "wandb",
    "huggingface_hub",
    "accelerate",
    "transformers",
    "lm-eval @ git+https://github.com/EleutherAI/lm-evaluation-harness.git",
]